+++ lib/seed_ui.dart
@@ -339,6 +339,8 @@ export 'src/theme/design_token.dart'
         SeedToken,
         SoftSize,
         Token;
+export 'src/theme/material_theme.dart'
+    show MaterialThemeFromThemeData, MaterialThemeFromToken;
 export 'src/theme/palette.dart' show generate;
 export 'src/utils/date_format.dart'
     show